imp: drop `smart` flag for boolean comparison
This is a breaking change.

In the end, there is no reason to use `B.or(a, b)` instead of `a || b` *except* if we do want the smart conversion to boolean. For that matter, it both simplifies the library and makes it easier to understand while aligning with the philosophy of primitive-kit: straightforward, simple and convenient.